fix auth.ts

+++ lib/auth.ts
@@ -2,18 +2,18 @@ import NextAuth from "next-auth";
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import { prisma } from "@/db/prisma";
 import CredentialsProvider from "next-auth/providers/credentials";
-import type { NextAuthConfig } from "next-auth";
 import { cookies } from "next/headers";
 import { compare } from "./encrypt";
 import { NextResponse } from "next/server";
-export const config = {
+
+export const { handlers, auth, signIn, signOut } = NextAuth({
   pages: {
     signIn: "/sign-in",
     error: "/sign-in",
   },
   session: {
     strategy: "jwt",
-    maxAge: 30 * 24 * 60 * 60, // 30 days
+    maxAge: 30 * 24 * 60 * 60,
   },
   adapter: PrismaAdapter(prisma),
   providers: [
@@ -22,27 +22,18 @@ export const config = {
         email: { type: "email" },
         password: { type: "password" },
       },
-
       async authorize(credentials) {
-        if (credentials == null) {
-          return null;
-        }
+        if (credentials == null) return null;
 
-        // Find user in database
         const user = await prisma.user.findFirst({
-          where: {
-            email: credentials.email as string,
-          },
+          where: { email: credentials.email as string },
         });
 
-        // Check if user exists and if the password matches
         if (user && user.password) {
           const isMatch = await compare(
             credentials.password as string,
             user.password,
           );
-
-          // If password is correct, return user
           if (isMatch) {
             return {
               id: user.id,
@@ -52,36 +43,27 @@ export const config = {
             };
           }
         }
-        // If user does not exist or password does not match return null
         return null;
       },
     }),
   ],
   callbacks: {
     async session({ session, user, trigger, token }: any) {
-      // Set the user ID from the token
       session.user.id = token.sub;
       session.user.role = token.role;
       session.user.name = token.name;
-
-      // If there is an update, set the user name
       if (trigger === "update") {
         session.user.name = user.name;
       }
-
       return session;
     },
     async jwt({ token, user, trigger, session }: any) {
-      // Assign user fields to token
       if (user) {
         token.id = user.id;
         token.role = user.role;
 
-        // If user has no name then use the email
         if (user.name === "NO_NAME") {
           token.name = user.email!.split("@")[0];
-
-          // Update database to reflect the token name
           await prisma.user.update({
             where: { id: user.id },
             data: { name: token.name },
@@ -98,12 +80,7 @@ export const config = {
             });
 
             if (sessionCart) {
-              // Delete current user cart
-              await prisma.cart.deleteMany({
-                where: { userId: user.id },
-              });
-
-              // Assign new cart
+              await prisma.cart.deleteMany({ where: { userId: user.id } });
               await prisma.cart.update({
                 where: { id: sessionCart.id },
                 data: { userId: user.id },
@@ -113,7 +90,6 @@ export const config = {
         }
       }
 
-      // Handle session updates
       if (session?.user.name && trigger === "update") {
         token.name = session.user.name;
       }
@@ -121,7 +97,6 @@ export const config = {
       return token;
     },
     authorized({ request, auth }: any) {
-      // Array of regex patterns of paths we want to protect
       const protectedPaths = [
         /\/shipping-address/,
         /\/payment-method/,
@@ -132,36 +107,20 @@ export const config = {
         /\/admin/,
       ];
 
-      // Get pathname from the req URL object
       const { pathname } = request.nextUrl;
 
-      // Check if user is not authenticated and accessing a protected path
       if (!auth && protectedPaths.some((p) => p.test(pathname))) return false;
 
-      // Check for session cart cookie
       if (!request.cookies.get("sessionCartId")) {
-        // Generate new session cart id cookie
         const sessionCartId = crypto.randomUUID();
-
-        // Clone the req ers
-        const newRequestHeaders = request.headers;
-
-        // Create new response and add the new ers
         const response = NextResponse.next({
-          request: {
-            headers: newRequestHeaders,
-          },
+          request: { headers: request.headers },
         });
-
-        // Set newly generated sessionCartId in the response cookies
         response.cookies.set("sessionCartId", sessionCartId);
-
         return response;
-      } else {
-        return true;
       }
+
+      return true;
     },
   },
-} satisfies NextAuthConfig;
-
-export const { handlers, auth, signIn, signOut } = NextAuth(config);
+});
